requirements.txt
asgiref==3.5.2
astroid==2.12.13
backports.zoneinfo==0.2.1
colorlog==6.7.0
dill==0.3.6
Django==4.1.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
platformdirs==2.6.0
pylint==2.15.8
python-dotenv==0.21.0
pytils==0.4.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
